requirements.txt
Django==4.2.7
mysqlclient==2.2.0
Pillow==10.1.0
python-decouple==3.8
gunicorn==21.2.0